backend/requirements.txt
# Backend Todo API Dependencies
# Feature: 001-backend-todo-api
# Python 3.11+

# Web Framework
fastapi>=0.100.0
uvicorn[standard]>=0.24.0

# Database ORM & Driver
sqlmodel>=0.0.14
asyncpg>=0.29.0

# Data Validation
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Environment Configuration
python-dotenv>=1.0.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
httpx>=0.25.0

# Authentication (JWT)
PyJWT[crypto]>=2.8.0
cachetools>=5.3.0
types-cachetools>=5.3.0
